Refactor Ext Mon Detector inside new mother volume
Resolve conflicts in rebase
