+++ Mu2eG4/geom/extmon_fnal_v02.txt
@@ -184,6 +184,14 @@ double         extMonFNAL.scintOffset = 29.0; // Distance from pixel plane cente
 double         extMonFNAL.scintInnerOffset = 33; // Distance from pixel plane center to center of scintillator on the magnet side
 double         extMonFNAL.scintGap = 5.3; // Distance between the surfaces of the two adjacent scintillators
 
+//================================================================
+// The Ext Mon detector mother
+
+vector<double> extMonFNAL.motherHalfSize = { 150, 200, 650 };
+double extMonFNAL.motherDistToMagnet = 1;
+bool extMonFNAL.stackRotation = false;
+
+//================================================================
 // The upstream stack of the sensors.
 //
 // Location in room: the center is on the continuation of
@@ -195,9 +203,6 @@ double         extMonFNAL.scintGap = 5.3; // Distance between the surfaces of th
 // Details of the detector.  All of these vectors should be
 // of the same size: one entry per detector plane.
 // The coordinates are relative to the detector box volume.
-vector<double> extMonFNAL.up.motherTransverseHalfSize = { 115, 115 }; // x, y
-double extMonFNAL.up.motherStartZ = 276.;
-double extMonFNAL.up.motherEndZ = 676.;
 
 vector<double> extMonFNAL.up.plane_zoffset = { +326., +426., +526., +626.};
 vector<double> extMonFNAL.up.plane_xoffset = { 0.,    0.,    0.,    0. };
@@ -222,8 +227,6 @@ vector<double> extMonFNAL.up.module_rotation = { 0., 0. };
 // The coordinates are relative to the detector box volume.
 
 vector<double> extMonFNAL.dn.motherTransverseHalfSize = { 115, 115 }; // x, y
-double extMonFNAL.dn.motherStartZ = -676.;
-double extMonFNAL.dn.motherEndZ = -276.;
 
 vector<double> extMonFNAL.dn.plane_zoffset = {-626., -526.,  -426., -326.};
 vector<double> extMonFNAL.dn.plane_xoffset = { 0., 0., 0., 0. };
